+++ CHANGELOG.md
@@ -5,7 +5,10 @@
   * status DSL
 
 * Deprecations
-  json/1, json/2, html/1, html/2, text/1, text/2 is deprecated in faver of returning data directily.
+  * json/1, json/2, html/1, html/2, text/1, text/2 is deprecated in faver of returning data directily.
+
+* Bugfix
+  * param value replaced by identical param keys in group [#5](https://github.com/falood/maru/issues/5).
 
 ## v0.3.0 (2015-6-2)
 * Enhancements